+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.18.14
 
+* Add support for decoding Document representations of untagged unions
 * Update aws-http4s clients using json to have a maxArity of Int.MaxValue
 
 # 0.18.13